remove ViewService interface
